requirements.txt
pytest==8.2.2
cryptography==43.0.1
PyJWT==2.9.0
requests==2.32.3
idna==3.7
